fitsbin: more 32- vs 64-bit large-file woes.  Fix bug report from Sergey.
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@22507 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/fitsbin.c
@@ -503,7 +503,7 @@ static int read_chunk(fitsbin_t* fb, fitsbin_chunk_t* chunk) {
         return -1;
     }
 
-    expected = chunk->itemsize * chunk->nrows;
+    expected = (size_t)chunk->itemsize * (size_t)chunk->nrows;
 	if (in_memory(fb)) {
 		int i;
 		chunk->data = malloc(expected);
